Resolve "US4: Configure Percentile calculations for the grid"

Closes #43 and #83

See merge request 2025ws-ase-pr-group/25ws-ase-pr-qse-05!53

+++ LiDARC-backend/src/main/java/com/example/lidarcbackend/model/DTO/StartPreProcessJobDto.java
@@ -22,8 +22,8 @@ public class StartPreProcessJobDto {
     private Long comparisonId;
     private Long fileId;
     private Double individualPercentile;
-    private Integer pointFilterLowerBound;
-    private Integer pointFilterUpperBound;
+    private Double pointFilterLowerBound;
+    private Double pointFilterUpperBound;
     private Boolean pointFilterEnabled;
     private Boolean outlierDetectionEnabled;
     private Double outlierDeviationFactor;
